Prevent Infinite While Loop
Created the isOperatorExist() condition that checks if there is an
operator in the current array. If there is, then the operator variable
takes the first available operator. If there isn't an operator, then the
operator variable value is unchanged.
This condition also checks if there is only one number remaining. If so,
then the input string is deleted by one extra character, to ensure that
the input array is empty, ending the while loop.